[CHORE] TodoItem 위치 변경
TodoItem의 위치를 common 패키지로 이동함.

Solves: TICO-44
Related to: TICO-33
