+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/ui/common/view/TodoItem.kt
@@ -1,0 +1,67 @@
+package com.tico.pomorodo.ui.common.view
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.unit.dp
+import com.tico.pomorodo.R
+import com.tico.pomorodo.data.model.TodoData
+import com.tico.pomorodo.data.model.TodoState
+import com.tico.pomorodo.ui.theme.IC_TODO_CHECKED
+import com.tico.pomorodo.ui.theme.IC_TODO_GOING
+import com.tico.pomorodo.ui.theme.IC_TODO_UNCHECKED
+import com.tico.pomorodo.ui.theme.PomoroDoTheme
+
+@Composable
+fun TodoItem(
+    iconSize: Int,
+    todoData: TodoData,
+    enabled: Boolean,
+    textStyle: TextStyle,
+    onStateChanged: ((TodoState) -> Unit)? = null
+) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(5.dp)
+    ) {
+        TodoCheckBox(
+            size = iconSize,
+            state = todoData.state,
+            enabled = enabled,
+            onStateChanged = { onStateChanged?.invoke(it) })
+        SimpleText(
+            modifier = Modifier,
+            text = todoData.name,
+            style = textStyle,
+            color = PomoroDoTheme.colorScheme.onBackground
+        )
+    }
+}
+
+@Composable
+private fun TodoCheckBox(
+    size: Int,
+    state: TodoState,
+    onStateChanged: (TodoState) -> Unit,
+    enabled: Boolean
+) {
+    SimpleIcon(
+        modifier = Modifier.clickableWithoutRipple(enabled) {
+            onStateChanged(state)
+        },
+        size = size,
+        imageVector = when (state) {
+            TodoState.UNCHECKED -> PomoroDoTheme.iconPack[IC_TODO_UNCHECKED]!!
+            TodoState.GOING -> PomoroDoTheme.iconPack[IC_TODO_GOING]!!
+            TodoState.CHECKED -> PomoroDoTheme.iconPack[IC_TODO_CHECKED]!!
+        },
+        contentDescriptionId = when (state) {
+            TodoState.UNCHECKED -> R.string.content_todo_unchecked
+            TodoState.GOING -> R.string.content_todo_going
+            TodoState.CHECKED -> R.string.content_todo_checked
+        }
+    )
+}
